fix: error code from uint8 to int32 in error class

+++ inc/osal/error.hpp
@@ -141,7 +141,7 @@ public:
      * @param func The function name where the error occurred.
      * @param line The line number where the error occurred.
      */
-    explicit error(const char* msg, uint8_t code = 0, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT;
+    explicit error(const char* msg, int32_t code = 0, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT;
 
     /**
      * @brief Constructs an error object with the given message, error type, file, function, and line number.
@@ -153,7 +153,7 @@ public:
      * @param line The line number where the error occurred.
      */
     explicit error(const char* msg, error_type code = error_type::OS_ENO, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT
-        : error(msg, static_cast<uint8_t>(code), file, func, line) {}
+        : error(msg, static_cast<int32_t>(code), file, func, line) {}
 
     /**
      * @brief Constructs an error object with the given previous error, message, code, file, function, and line number.
@@ -165,7 +165,7 @@ public:
      * @param func The function name where the error occurred.
      * @param line The line number where the error occurred.
      */
-    error(error* old_error, const char* msg, uint8_t code = 0, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT;
+    error(error* old_error, const char* msg, int32_t code = 0, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT;
 
     /**
      * @brief Constructs an error object with the given previous error, message, code, file, function, and line number.
@@ -178,7 +178,7 @@ public:
      * @param line The line number where the error occurred.
      */
     error(error* old_error, const char* msg, error_type code = error_type::OS_ENO, const char* file = get_file_name(__FILE__), const char* func = "", uint32_t line = __LINE__) OSAL_NOEXCEPT
-    : error(old_error, msg, static_cast<uint8_t>(code), file, func, line)
+    : error(old_error, msg, static_cast<int32_t>(code), file, func, line)
     {
 
     }
@@ -233,7 +233,7 @@ public:
      *
      * @return The error code.
      */
-    inline uint8_t get_code() const OSAL_NOEXCEPT
+    inline uint32_t get_code() const OSAL_NOEXCEPT
     {
         return code;
     }
